Updated for Directory check of that specific date
*As per the directions of the Project the CSV file with particular incremental number will be saved in a directory named in that the then date.
Here we run a for loop which checks the directory listing and checks whether the directory with that current date exists. when the direcory exist there is a
variable count which provides itself as a counter for the check and stops the loop when the value of it get to 1.

*Certain specific cleanup with gitignore is being done(Apology for dirtying this commit)